+++ vm/src/stdlib/posix.rs
@@ -23,7 +23,7 @@ pub mod module {
     use crate::{
         builtins::{PyDictRef, PyInt, PyListRef, PyStrRef, PyTupleRef, PyTypeRef},
         convert::{IntoPyException, ToPyObject, TryFromObject},
-        function::{Either, OptionalArg},
+        function::{Either, OptionalArg, KwArgs},
         stdlib::os::{
             errno_err, DirFd, FollowSymlinks, OsPath, OsPathOrFd, SupportFunc, TargetIsDirectory,
             _os, fs_metadata, IOErrorBuilder,
@@ -419,6 +419,122 @@ pub mod module {
             FollowSymlinks(true),
             vm,
         )
+    }
+
+    #[derive(FromArgs)]
+    struct RegisterAtForkArgs {
+        #[pyarg(named, optional)]
+        before: OptionalArg<PyObjectRef>,
+        #[pyarg(named, optional)]
+        after_in_parent: OptionalArg<PyObjectRef>,
+        #[pyarg(named, optional)]
+        after_in_child: OptionalArg<PyObjectRef>,
+    }
+
+    impl RegisterAtForkArgs {
+        fn into_validated(
+            self,
+            vm: &VirtualMachine,
+        ) -> PyResult<(
+            Option<PyObjectRef>,
+            Option<PyObjectRef>,
+            Option<PyObjectRef>,
+        )> {
+            fn into_option(
+                arg: OptionalArg<PyObjectRef>,
+                vm: &VirtualMachine,
+            ) -> PyResult<Option<PyObjectRef>> {
+                match arg {
+                    OptionalArg::Present(obj) => {
+                        if !obj.is_callable() {
+                            return Err(vm.new_type_error("Args must be callable".to_owned()));
+                        }
+                        Ok(Some(obj))
+                    }
+                    OptionalArg::Missing => Ok(None),
+                }
+            }
+            let before = into_option(self.before, vm)?;
+            let after_in_parent = into_option(self.after_in_parent, vm)?;
+            let after_in_child = into_option(self.after_in_child, vm)?;
+            if before.is_none() && after_in_parent.is_none() && after_in_child.is_none() {
+                return Err(vm.new_type_error("At least one arg must be present".to_owned()));
+            }
+            Ok((before, after_in_parent, after_in_child))
+        }
+    }
+
+    #[pyfunction]
+    fn register_at_fork(
+        args: RegisterAtForkArgs,
+        _ignored: KwArgs,
+        vm: &VirtualMachine,
+    ) -> PyResult<()> {
+        let (before, after_in_parent, after_in_child) = args.into_validated(vm)?;
+
+        if let Some(before) = before {
+            vm.state.before_forkers.lock().push(before);
+        }
+        if let Some(after_in_parent) = after_in_parent {
+            vm.state.after_forkers_parent.lock().push(after_in_parent);
+        }
+        if let Some(after_in_child) = after_in_child {
+            vm.state.after_forkers_child.lock().push(after_in_child);
+        }
+        Ok(())
+    }
+
+    fn run_at_forkers(mut funcs: Vec<PyObjectRef>, reversed: bool, vm: &VirtualMachine) {
+        if !funcs.is_empty() {
+            if reversed {
+                funcs.reverse();
+            }
+            for func in funcs.into_iter() {
+                if let Err(e) = func.call((), vm) {
+                    let exit = e.fast_isinstance(vm.ctx.exceptions.system_exit);
+                    vm.run_unraisable(e, Some("Exception ignored in".to_owned()), func);
+                    if exit {
+                        // Do nothing!
+                    }
+                }
+            }
+        }
+    }
+
+    fn py_os_before_fork(vm: &VirtualMachine) -> PyResult<()> {
+        let before_forkers: Vec<PyObjectRef> = vm.state.before_forkers.lock().clone();
+        // functions must be executed in reversed order as they are registered
+        // only for before_forkers, refer: test_register_at_fork in test_posix
+
+        run_at_forkers(before_forkers, true, vm);
+        Ok(())
+    }
+
+    fn py_os_after_fork_child(vm: &VirtualMachine) -> PyResult<()> {
+        let after_forkers_child: Vec<PyObjectRef> = vm.state.after_forkers_child.lock().clone();
+        run_at_forkers(after_forkers_child, false, vm);
+        Ok(())
+    }
+
+    fn py_os_after_fork_parent(vm: &VirtualMachine) -> PyResult<()> {
+        let after_forkers_parent: Vec<PyObjectRef> = vm.state.after_forkers_parent.lock().clone();
+        run_at_forkers(after_forkers_parent, false, vm);
+        Ok(())
+    }
+
+    #[pyfunction]
+    fn fork(vm: &VirtualMachine) -> PyResult {
+        let pid: i32;
+        py_os_before_fork(vm)?;
+        unsafe {
+            pid = libc::fork();
+        }
+        if pid == 0 {
+            py_os_after_fork_child(vm)?;
+        } else {
+            py_os_after_fork_parent(vm)?;
+        }
+        Ok(vm.ctx.new_int(pid).into())
     }
 
     #[cfg(not(target_os = "redox"))]
